requirements.txt
google-generativeai
requests
pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
pyyaml
python-dotenv
tenacity
pathlib
